\name{arc_radius_locs}
\alias{arc_radius_locs}
\title{Arc Radius Locations}
\usage{
  arc_radius_locs(edgelist, nodes, centers)
}
\arguments{
  \item{edgelist}{2-column matrix}

  \item{nodes}{vector of nodes}

  \item{centers}{vector with xy-positions of nodes}
}
\value{
  a list with locations and radios

  \item{locs}{locations}

  \item{radios}{radius values}
}
\description{
  Computes the location and radius of each arc
}
\keyword{internal}

