\name{min_max_margin}
\alias{min_max_margin}
\title{Minimum and Maximum Margin Limits}
\usage{
  min_max_margin(radios, above)
}
\arguments{
  \item{radios}{vector of arc radius}

  \item{above}{logical vectors indicating whether arcs
  should be displayed}
}
\value{
  list with minimum and maximum margin limits
}
\description{
  Computes the minimum and maximum margin limits of
  plotting region
}
\keyword{internal}

