% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcplot.r
\name{xynodes}
\alias{xynodes}
\title{X or Y coordinates of node locations}
\usage{
xynodes(num_nodes, aux_ord, labels)
}
\arguments{
\item{num_nodes}{number of nodes}

\item{aux_ord}{vector with the index number for ordering the nodes}

\item{labels}{optional string vector with labels for the nodes}
}
\description{
Gives axis locations of each node
}
